requirements.txt
torch
langchain
ipykernel
openai
streamlit==1.24.0
python-env
PyPDF2
pandas
transformers==4.30
pytesseract
opencv-python-headless
pillow
imagehash
pytube
ffmpeg
openai-whisper
bitsandbytes
bitsandbytes-cuda117
accelerate
